Move config files to a new dir
