+++ src/__init__.py
@@ -1,0 +1,7 @@
+#-----------------------------------------------------------------
+# policychecker: __init__.py
+#
+# This package file exports some convenience functions for
+# interacting with policychecker
+#
+#-----------------------------------------------------------------
